Use the heap more strategically to minimize the memory footprint as we iterate through files
Use a MaxHeap (pop-order is newest access time first). Populate the heap
until we have enough files in it to reach our target capacity. From then
on, as we add files to the heap, pop the most-recently-accessed ones
that we no longer need to clean up to reach our target.